+++ modules/os2forms_encrypt/src/Form/SettingsForm.php
@@ -3,6 +3,7 @@
 namespace Drupal\os2forms_encrypt\Form;
 
 use Drupal\Core\Config\ConfigFactoryInterface;
+use Drupal\Core\Config\TypedConfigManagerInterface;
 use Drupal\Core\Form\ConfigFormBase;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\Core\Link;
@@ -28,10 +29,10 @@ class SettingsForm extends ConfigFormBase {
    *
    * @var \Drupal\encrypt\EncryptionProfileManager
    */
-  private EncryptionProfileManager $encryptionProfileManager;
+  protected EncryptionProfileManager $encryptionProfileManager;
 
-  public function __construct(ConfigFactoryInterface $config_factory, EncryptionProfileManager $encryptionProfileManager) {
-    parent::__construct($config_factory);
+  public function __construct(ConfigFactoryInterface $config_factory, TypedConfigManagerInterface $typedConfigManager, EncryptionProfileManager $encryptionProfileManager) {
+    parent::__construct($config_factory, $typedConfigManager);
     $this->encryptionProfileManager = $encryptionProfileManager;
   }
 
@@ -41,6 +42,7 @@ class SettingsForm extends ConfigFormBase {
   public static function create(ContainerInterface $container) {
     return new static(
       $container->get('config.factory'),
+      $container->get('config.typed'),
       $container->get('encrypt.encryption_profile.manager')
     );
   }