+++ modules/os2forms_permissions_by_term/src/Form/SettingsForm.php
@@ -24,7 +24,7 @@ class SettingsForm extends ConfigFormBase {
    *
    * @var \Drupal\Core\Entity\EntityFieldManagerInterface
    */
-  private $entityFieldTypeManager;
+  protected $entityFieldTypeManager;
 
   /**
    * Constructs an SettingsForm object.